Deprecate WorldComponent
