+++ blockycraft/Assets/Scripts/Behaviours/Player.cs
@@ -1,5 +1,4 @@
 ﻿using Blockycraft;
-using Blockycraft.World;
 using UnityEngine;
 
 public sealed class Player : MonoBehaviour
@@ -40,9 +39,9 @@ public sealed class Player : MonoBehaviour
     private Vector3Int GetChunkCoordFromPosition(Vector3 position)
     {
         return new Vector3Int(
-            (int)(position.x / WorldComponent.SIZE),
-            (int)(position.y / WorldComponent.SIZE),
-            (int)(position.z / WorldComponent.SIZE)
+            (int)(position.x / World.SIZE),
+            (int)(position.y / World.SIZE),
+            (int)(position.z / World.SIZE)
         );
     }
