+++ blockycraft/Assets/Scripts/Behaviours/World.cs
@@ -6,29 +6,67 @@ using UnityEngine;
 
 public sealed class World : MonoBehaviour
 {
+    public const int SIZE = 8;
     public const int DRAW_HEIGHT = 4;
     public const int DRAW_DISTANCE = DRAW_HEIGHT * 2;
-    public WorldComponent component;
     public Material material;
-    public ChunkGenerator start;
+
     private System3D<Chunk> chunks;
+    public System3D<ChunkBlocks> Chunks { get; private set; }
     private ChunkFactory factory;
+    private int radius;
+
+    public ChunkGenerator start;
+    private ChunkGenerator current;
 
     public void Ping(Vector3Int center)
     {
-        component.Ping(center);
+        var biome = current;
+        if (Random.value < 0.40f)
+        {
+            biome = current.Transitions[(int)(Random.value * (current.Transitions.Length))];
+            current = biome;
+        }
+
+        Chunks.Ping(center, radius, v =>
+        {
+            var chunk = new ChunkBlocks(v.x, v.y, v.z, SIZE);
+            return biome.Generate(v, chunk, SIZE);
+        });
+
         chunks.Ping(center, DRAW_HEIGHT, (v) =>
         {
-            var blocks = component.Chunks.Get(v.x, v.y, v.z);
+            var blocks = Chunks.Get(v.x, v.y, v.z);
             var mesh = ChunkFactory.Build(blocks);
             return Chunk.Create(blocks, material, v.x, v.y, v.z, gameObject, mesh);
         });
     }
 
+    public BlockType GetBlock(int x, int y, int z)
+    {
+        var coord = MathHelper.Anchor(x, y, z, SIZE);
+        var block = MathHelper.Wrap(x, y, z, SIZE);
+        if (!Chunks.TryGet(ref coord, out ChunkBlocks blocks))
+        {
+            return null;
+        }
+
+        if (!blocks.TryGet(ref block, out BlockType type))
+        {
+            return null;
+        }
+        return type;
+    }
+
+    public static string Key(int x, int y, int z)
+    {
+        return $"{x}:{y}:{z}";
+    }
+
     public void Set(Vector3 target, BlockType type)
     {
-        var coord = MathHelper.Anchor(Mathf.FloorToInt(target.x), Mathf.FloorToInt(target.y), Mathf.FloorToInt(target.z), WorldComponent.SIZE);
-        var block = MathHelper.Wrap(Mathf.FloorToInt(target.x), Mathf.FloorToInt(target.y), Mathf.FloorToInt(target.z), WorldComponent.SIZE);
+        var coord = MathHelper.Anchor(Mathf.FloorToInt(target.x), Mathf.FloorToInt(target.y), Mathf.FloorToInt(target.z), SIZE);
+        var block = MathHelper.Wrap(Mathf.FloorToInt(target.x), Mathf.FloorToInt(target.y), Mathf.FloorToInt(target.z), SIZE);
         if (!chunks.TryGet(ref coord, out Chunk chunk))
         {
             return;
@@ -44,7 +82,7 @@ public sealed class World : MonoBehaviour
         while (step < reach)
         {
             Vector3 pos = position + (forward * step);
-            var type = component.GetBlock(Mathf.FloorToInt(pos.x), Mathf.FloorToInt(pos.y), Mathf.FloorToInt(pos.z));
+            var type = GetBlock(Mathf.FloorToInt(pos.x), Mathf.FloorToInt(pos.y), Mathf.FloorToInt(pos.z));
             if (type != null && type.isVisible)
             {
                 pos = new Vector3(Mathf.FloorToInt(pos.x), Mathf.FloorToInt(pos.y), Mathf.FloorToInt(pos.z));
@@ -58,9 +96,11 @@ public sealed class World : MonoBehaviour
 
     private void Start()
     {
-        component = new WorldComponent(DRAW_DISTANCE * 2, start);
+        radius = DRAW_DISTANCE;
         chunks = new System3D<Chunk>();
+        Chunks = new System3D<ChunkBlocks>();
         factory = new ChunkFactory();
+        current = start;
 
         Ping(Vector3Int.zero);
     }
