--- blockycraft/Assets/Scripts/World.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 847bda8fb8ed27c45b139eb117fc76c5
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
