--- blockycraft/Assets/Scripts/World/Chunk.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 61ec375844ba85e4ea7709c30c9daf48
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
