--- blockycraft/Assets/Scripts/World/WorldComponent.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8da880670fe40584b9505ee24140841c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
